See https://github.com/nova-video-player/aos-AVP/issues/169
The problem is that in db onCreate von_insert table is not created which on certain custom ROMs creates a systematic crash at startup when trying to set the triggers.

This bug was introduced in db version 36.

Hopefully this fixes the issue. Keeping the issue open until proper confirmation (I cannot reproduce).